update categories api

+++ src/controllers/categoryControllers.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import asyncHandler from "express-async-handler";
 import { validationResult } from "express-validator";
 import Business from "../models/BusinessModel.js";
@@ -12,6 +13,12 @@ const getCategories = asyncHandler(async (req, res) => {
   }
 
   const business_id = req.query.business_id;
+
+  if (!mongoose.Types.ObjectId.isValid(business_id)) {
+    return res.status(400).json({
+      messsage: "Invalid business id",
+    });
+  }
 
   const business = await Business.findById(business_id).populate("categories");
 
@@ -33,6 +40,12 @@ const createCategory = asyncHandler(async (req, res) => {
   }
 
   const business_id = req.body.business_id;
+
+  if (!mongoose.Types.ObjectId.isValid(business_id)) {
+    return res.status(400).json({
+      messsage: "Invalid business id",
+    });
+  }
 
   const business = await Business.findById(business_id);
 
@@ -65,6 +78,12 @@ const getCategory = asyncHandler(async (req, res) => {
 
   const category_id = req.params.category_id;
 
+  if (!mongoose.Types.ObjectId.isValid(category_id)) {
+    return res.status(400).json({
+      messsage: "Invalid category id",
+    });
+  }
+
   const category = await Category.findById(category_id).populate("products");
 
   if (category) {
@@ -84,7 +103,13 @@ const updateCategory = asyncHandler(async (req, res) => {
     return res.json({ errors: errors.array() });
   }
 
-  const category_id = req.body.category_id;
+  const category_id = req.params.category_id;
+
+  if (!mongoose.Types.ObjectId.isValid(category_id)) {
+    return res.status(400).json({
+      messsage: "Invalid category id",
+    });
+  }
 
   const category = await Category.findById(category_id);
 
@@ -112,6 +137,12 @@ const deleteCategory = asyncHandler(async (req, res) => {
   }
 
   const category_id = req.params.category_id;
+
+  if (!mongoose.Types.ObjectId.isValid(category_id)) {
+    return res.status(400).json({
+      messsage: "Invalid category id",
+    });
+  }
 
   const category = await Category.findByIdAndDelete(category_id);
